requirements.txt
Appium_Python_Client==4.0.1
behave==1.2.6
selenium==4.23.0
Paver==1.3.4
allure-behave==2.13.5
allure_python_commons==2.13.5